@@ -7,18 +7,15 @@ namespace App\User\Command;
 use Violines\RestBundle\HttpApi\HttpApi;
 use Symfony\Component\Validator\Constraints as Assert;
 
-/**
- * @HttpApi
- */
+#[HttpApi]
 class CreateProfile
 {
-    /**
-     * @Assert\Type("string")
-     */
+    #[Assert\Type("string")]
+    #[Assert\Email]
+    #[Assert\NotNull]
     public $email;
 
-    /**
-     * @Assert\Type("string")
-     */
+    #[Assert\Type("string")]
+    #[Assert\NotNull]
     public $password;
 }
