@@ -8,29 +8,19 @@ use Violines\RestBundle\HttpApi\HttpApi;
 use Symfony\Component\Serializer;
 use Symfony\Component\Validator\Constraints as Assert;
 
-/**
- * @HttpApi
- */
+#[HttpApi]
 class EditProfile
 {
-    /**
-     * @Assert\Type("int")
-     * @Serializer\Annotation\SerializedName("user_id")
-     */
+    #[Assert\Type("int")]
+    #[Serializer\Annotation\SerializedName("user_id")]
     public $userId;
 
-    /**
-     * @Assert\Type("string")
-     */
+    #[Assert\Type("string")]
     public $email;
 
-    /**
-     * @Assert\Type("boolean")
-     */
+    #[Assert\Type("boolean")]
     public $isResetPassword;
 
-    /**
-     * @Assert\Type("boolean")
-     */
+    #[Assert\Type("boolean")]
     public $isResetKey;
 }
